extract Button styles
