+++ src/components/atoms/IconChip/IconChip.tsx
@@ -1,30 +1,13 @@
 import React from 'react';
 import { Chip, ChipProps } from '@mui/material';
+import styles from './IconChipStyles';
 
 interface IconChipProps extends Omit<ChipProps, 'sx'> {
   gradient?: boolean;
 }
 
 const IconChip: React.FC<IconChipProps> = ({ gradient = false, ...props }) => {
-  const sx = gradient
-    ? {
-        background: 'linear-gradient(135deg, #0f0fcf, #00d4ff)',
-        color: 'white',
-        fontWeight: 'bold',
-        fontSize: '1rem',
-        py: 2.5,
-        px: 2,
-        transition: 'transform 0.2s',
-        '&:hover': {
-          transform: 'scale(1.05)'
-        }
-      }
-    : {
-        bgcolor: 'rgba(255,255,255,0.2)',
-        color: 'white',
-        fontSize: '1rem',
-        py: 2.5
-      };
+  const sx = gradient ? styles.gradientChip : styles.defaultChip;
 
   return <Chip sx={sx} {...props} />;
 };
