+++ src/components/atoms/SecondaryButton/SecondaryButton.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Button, ButtonProps } from '@mui/material';
+import styles from './SecondaryButtonStyles';
 
 interface SecondaryButtonProps extends ButtonProps {
   light?: boolean;
@@ -11,35 +12,9 @@ const SecondaryButton: React.FC<SecondaryButtonProps> = ({
   sx = {},
   ...props
 }) => {
-  const baseStyles = light ? {
-    borderColor: 'white',
-    color: 'white',
-    fontWeight: 'bold',
-    px: 4,
-    py: 1.5,
-    borderWidth: 2,
-    '&:hover': {
-      borderColor: 'white',
-      borderWidth: 2,
-      bgcolor: 'rgba(255,255,255,0.1)',
-      transform: 'translateY(-2px)'
-    },
-    ...sx
-  } : {
-    borderColor: '#0f0fcf',
-    color: '#0f0fcf',
-    fontWeight: 'bold',
-    px: 4,
-    py: 1.5,
-    borderWidth: 2,
-    '&:hover': {
-      borderColor: '#0f0fcf',
-      borderWidth: 2,
-      bgcolor: 'rgba(15,15,207,0.05)',
-      transform: 'translateY(-2px)'
-    },
-    ...sx
-  };
+  const baseStyles = light ?
+    { ...styles.lightButton, ...sx } :
+    { ...styles.darkButton, ...sx };
 
   return (
     <Button
