+++ src/components/atoms/SecondaryButton/SecondaryButtonStyles.ts
@@ -1,0 +1,33 @@
+import { SxProps } from "@mui/material";
+
+const lightButton: SxProps = {
+  borderColor: 'white',
+  color: 'white',
+  fontWeight: 'bold',
+  px: 4,
+  py: 1.5,
+  borderWidth: 2,
+  '&:hover': {
+    borderColor: 'white',
+    borderWidth: 2,
+    bgcolor: 'rgba(255,255,255,0.1)',
+    transform: 'translateY(-2px)'
+  }
+};
+
+const darkButton: SxProps = {
+  borderColor: '#0f0fcf',
+  color: '#0f0fcf',
+  fontWeight: 'bold',
+  px: 4,
+  py: 1.5,
+  borderWidth: 2,
+  '&:hover': {
+    borderColor: '#0f0fcf',
+    borderWidth: 2,
+    bgcolor: 'rgba(15,15,207,0.05)',
+    transform: 'translateY(-2px)'
+  }
+};
+
+export default { lightButton, darkButton };
