+++ src/components/atoms/StatCard/StatCardStyles.ts
@@ -1,0 +1,12 @@
+import { SxProps } from "@mui/material";
+
+const valueText: SxProps = {
+  fontWeight: 'bold',
+  color: 'white'
+};
+
+const labelText: SxProps = {
+  color: 'rgba(255,255,255,0.8)'
+};
+
+export default { valueText, labelText };
