+++ src/components/atoms/TextButton/TextButton.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { Button, ButtonProps } from '@mui/material';
+import styles from './TextButtonStyles';
 
 interface TextButtonProps extends ButtonProps {
   light?: boolean;
@@ -11,27 +12,9 @@ const TextButton: React.FC<TextButtonProps> = ({
   sx = {},
   ...props
 }) => {
-  const baseStyles = light ? {
-    color: 'white',
-    fontWeight: 'bold',
-    px: 4,
-    py: 1.5,
-    '&:hover': {
-      bgcolor: 'rgba(255,255,255,0.1)',
-      transform: 'translateY(-2px)'
-    },
-    ...sx
-  } : {
-    color: '#0f0fcf',
-    fontWeight: 'bold',
-    px: 4,
-    py: 1.5,
-    '&:hover': {
-      bgcolor: 'rgba(15,15,207,0.05)',
-      transform: 'translateY(-2px)'
-    },
-    ...sx
-  };
+  const baseStyles = light ?
+    { ...styles.lightTextButton, ...sx } :
+    { ...styles.darkTextButton, ...sx };
 
   return (
     <Button
